[refactor] : foodItem -> mockItem 으로 변경
