+++ src/main.js
@@ -3,7 +3,7 @@ import { FoodItem } from "./component/FoodItem.js";
 import { FoodList } from "./component/FoodList.js";
 import { HeaderComponent } from "./component/HeaderComponent.js";
 import { Modal } from "./component/Modal.js";
-import { foodItems } from "./mock/foodItems.js";
+import { foodItems } from "./mock/mockItem.js";
 import { FoodListPage } from "./pages/FoodListPage.js";
 
 addEventListener("load", () => {
