+++ src/pages/FoodListPage.js
@@ -2,7 +2,7 @@ import { FoodForm } from "../component/FoodForm.js";
 import { FoodList } from "../component/FoodList.js";
 import { HeaderComponent } from "../component/HeaderComponent.js";
 import { Modal } from "../component/Modal.js";
-import { foodItems } from "../mock/foodItems.js";
+import { foodItems } from "../mock/mockItem.js";
 import { modalClose, modalOpen } from "../util/modalAction.js";
 
 function init() {
